# backend/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "API Documentation",
    "description": "Documentação da API do Project_paw",
    "contact": {}
  },
  "host": "localhost:3000",
  "basePath": "/",
  "securityDefinitions": {
    "JWT": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. Example: 'Bearer {token}'"
    }
  },
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
...
  "paths": {
...
    "/api/orders/ordersHistory": {
      "get": {
        "description": "Lista o histórico de encomendas do utilizador autenticado.",
        "summary": "Histórico de encomendas",
        "tags": [
          "Encomendas"
        ],
        "operationId": "HistoricoEncomendas",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de encomendas",
            "schema": {
              "type": "array",
              "items": {
                "type": "object"
              }
            }
          },
          "403": {
            "description": "Acesso negado"
          }
        }
      }
    },
    "/api/orders/hasOrder": {
      "get": {
        "description": "Verifica se o utilizador já fez encomenda num restaurante.",
        "summary": "Verificar encomenda em restaurante",
        "tags": [
          "Encomendas"
        ],
        "operationId": "VerificarEncomendaRestaurante",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "restaurantId",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "ID do restaurante"
          }
        ],
        "responses": {
          "200": {
            "description": "Resultado da verificação",
            "schema": {
              "type": "object",
              "properties": {
                "hasOrder": {
                  "type": "boolean"
                }
              }
            }
          },
          "403": {
            "description": "Acesso negado"
          }
        }
      }
    },
    "/api/orders/{orderId}": {
      "get": {
        "description": "Obtém os detalhes de uma encomenda pelo ID.",
        "summary": "Detalhes da encomenda",
        "tags": [
          "Encomendas"
        ],
        "operationId": "DetalhesEncomenda",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID da encomenda"
          }
        ],
        "responses": {
          "200": {
            "description": "Detalhes da encomenda",
            "schema": {
              "type": "object"
            }
          },
          "404": {
            "description": "Encomenda não encontrada"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/orders/{orderId}/delete": {
      "delete": {
        "description": "Remove uma encomenda pelo ID.",
        "summary": "Apagar encomenda",
        "tags": [
          "Encomendas"
        ],
        "operationId": "ApagarEncomenda",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID da encomenda"
          }
        ],
        "responses": {
          "200": {
            "description": "Encomenda apagada"
          },
          "404": {
            "description": "Encomenda não encontrada"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/orders/{orderId}/status": {
      "patch": {
        "description": "Atualiza o status de uma encomenda.",
        "summary": "Atualizar status da encomenda",
        "tags": [
          "Encomendas"
        ],
        "operationId": "AtualizarStatusEncomenda",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID da encomenda"
          },
          {
            "name": "status",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Status atualizado"
          },
          "404": {
            "description": "Encomenda não encontrada"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    }
  },
...
  "security": [],
...
}